[tls] Separate interface for handshaking
